+++ internal/handshake/tls_extension_handler_client.go
@@ -3,7 +3,6 @@ package handshake
 import (
 	"errors"
 	"fmt"
-	"math"
 
 	"github.com/lucas-clemente/quic-go/qerr"
 
@@ -121,8 +120,6 @@ func (h *extensionHandlerClient) Receive(hType mint.HandshakeType, el *mint.Exte
 	if err != nil {
 		return err
 	}
-	// TODO(#878): remove this when implementing the MAX_STREAM_ID frame
-	params.MaxStreams = math.MaxUint32
 	h.paramsChan <- *params
 	return nil
 }